Move App Bridge & Polaris web component into the app/ and login routes respectively.
This fixes an issue where App Bridge & App Bridge components are loaded on the marketing index page.  App Bridge especially should not be loaded on the marketing page, because it is not embedded
