+++ app/routes/auth.login/route.tsx
@@ -27,22 +27,25 @@ export default function Auth() {
   const { errors } = actionData || loaderData;
 
   return (
-    <s-page>
-      <Form method="post">
-        <s-section heading="Log in">
-          <s-text-field
-            type="text"
-            name="shop"
-            label="Shop domain"
-            helpText="example.myshopify.com"
-            value={shop}
-            onChange={setShop}
-            autoComplete="on"
-            error={errors.shop}
-          ></s-text-field>
-          <s-button submit>Log in</s-button>
-        </s-section>
-      </Form>
-    </s-page>
+    <>
+      <script src="https://cdn.shopify.com/shopifycloud/app-bridge-ui-experimental.js"></script>
+      <s-page>
+        <Form method="post">
+          <s-section heading="Log in">
+            <s-text-field
+              type="text"
+              name="shop"
+              label="Shop domain"
+              helpText="example.myshopify.com"
+              value={shop}
+              onChange={setShop}
+              autoComplete="on"
+              error={errors.shop}
+            ></s-text-field>
+            <s-button submit>Log in</s-button>
+          </s-section>
+        </Form>
+      </s-page>
+    </>
   );
 }
